requirements.txt
flask==2.3.3
opencv-python==4.8.0.74
numpy==1.24.3
ultralytics==8.0.145
python-dotenv==1.0.0
firebase-admin==6.2.0
pillow==10.0.0
flask-cors==4.0.0
